Refactor wiretrace logic into separate files
